add EstateId header constant
